0206 jihye_chat
- chat DB (txt 파일) 연결 작업
- chat.html, chat.js, chat 컨트롤러들 수정 필요
