+++ src/main/resources/static/js/script.js
@@ -162,7 +162,7 @@ function search_close() {
 // 채팅 기능에 사용 중 - 상단바 채팅버튼 & 부끄마켓 상세페이지창 채팅하기버튼
     function openWindow(url) {
         ChatWindow = window.open("about:blank", "Chat", 
-            "width=1200,height=700,left=350,top=150,dependent=yes,location=no, menubar=no,status=no,resizable=no,toolbar=no,scrollbars=yes");
+            "width=1200,height=700,left=350,top=150,dependent=yes,location=yes, menubar=no,status=no,resizable=no,toolbar=no,scrollbars=yes");
         ChatWindow.location.href = url;
         ChatWindow.focus();
         return false;
