+++ src/main/resources/templates/chat.html
@@ -72,74 +72,35 @@
 
 <body>
 
-    <div class="container p-0">
-
-        <h1 class="h3 mb-3">Messages</h1>
-
-        <div class="card"><!-- 내 채팅목록 & 현재 채팅 화면 -->
-            <div class="row g-0">
-                <div class="col-12 col-lg-5 col-xl-3 border-right">
-
-                    <div class="px-4 d-none d-md-block">
-                        <div class="d-flex align-items-center">
-                            <div class="flex-grow-1">
-                                <input type="hidden" id="loginUser" th:value="${#authentication.name}" />
-                                <input type="hidden" id="startChatWithYou" th:value="${ firstConnectUser }" />
-                                <input id="chatUser" type="text" class="form-control my-3" placeholder="Search...">
-                                <button id="refresh" class="form-control my-3">Refresh</button>
-                            </div>
-                        </div>
-                    </div>
-
-                    <!-- 내 대화 목록 -->
-                    <div id="myList">
-                    <ul class="list">
-
-                    </ul>
-                    </div>
-                    <hr class="d-block d-lg-none mt-1 mb-0">
-                </div>
-                <div class="col-12 col-lg-7 col-xl-9"><!-- 현재 채팅 화면(오른쪽 화면) -->
-                    <div class="py-2 px-4 border-bottom d-none d-lg-block">
-                        <div class="d-flex align-items-center py-1">
-                            <div class="position-relative">
-                                <img src="https://bootdey.com/img/Content/avatar/avatar3.png" class="rounded-circle mr-1" alt="Sharon Lessman" width="40" height="40">
-                            </div>
-                            <div class="flex-grow-1 pl-3">
-                                <strong id="selectedUserId"></strong>
-                                <div class="text-muted small"><em>Typing...</em></div>
-                            </div>
-                            <div>
-                                <button class="btn btn-primary btn-lg mr-1 px-3"><svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-phone feather-lg"><path d="M22 16.92v3a2 2 0 0 1-2.18 2 19.79 19.79 0 0 1-8.63-3.07 19.5 19.5 0 0 1-6-6 19.79 19.79 0 0 1-3.07-8.67A2 2 0 0 1 4.11 2h3a2 2 0 0 1 2 1.72 12.84 12.84 0 0 0 .7 2.81 2 2 0 0 1-.45 2.11L8.09 9.91a16 16 0 0 0 6 6l1.27-1.27a2 2 0 0 1 2.11-.45 12.84 12.84 0 0 0 2.81.7A2 2 0 0 1 22 16.92z"></path></svg></button>
-                                <button class="btn btn-info btn-lg mr-1 px-3 d-none d-md-inline-block"><svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-video feather-lg"><polygon points="23 7 16 12 23 17 23 7"></polygon><rect x="1" y="5" width="15" height="14" rx="2" ry="2"></rect></svg></button>
-                                <button class="btn btn-light border btn-lg px-3"><svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-more-horizontal feather-lg"><circle cx="12" cy="12" r="1"></circle><circle cx="19" cy="12" r="1"></circle><circle cx="5" cy="12" r="1"></circle></svg></button>
-                            </div>
-                        </div>
-                    </div>
-
-                    <!-- 채팅 내용 들어갈 공간 -->
-                    <div class="chat-history">
-                        <ul>
-
-                        </ul>
-                    </div>
-                    
-                    <!-- 채팅 작성 형식란 -->
-                    <div class="flex-grow-0 py-3 px-4 border-top">
-                        <div class="input-group">
-                            <input id="chat-to-send" type="text" class="form-control" placeholder="보낼 내용을 입력하세요.">
-                            <button id="btnSend" class="btn btn-primary">보내기</button>
-                        </div>
-                    </div><!-- 채팅 작성 형식란 끝 -->
-
-                </div><!-- 현재 채팅 화면(오른쪽 화면) 끝 -->
+    <div class="container">
+            <div class="title_text">
+                <h2>Booque Chat</h2>
             </div>
-        </div><!-- 내 채팅목록 & 현재 채팅 화면 끝 -->
-    </div><!-- 컨테이너 끝 -->
-    
-    
-    
-    </div>
+            <div class="row">
+                    <!-- chatHistory와 member가 실시간 입력하는 메시지 출력 -->
+                    <div id="content">
+                        <div th:each="chat : ${chatHistory}">
+                            <p>
+                                <span id="chatSender">${chat.sender}</span><br>
+                                <span id="chatMessage">${chat.message}</span><br>
+                                <span id="chatSendTime">${chat.sendTime}</span><br>
+                            </p>    
+                        </div>
+                    </div>
+                    <!-- 메시지 입력창과 보내기 버튼 -->
+                    <div class="row_3">
+                        <div class="input_group" id="sendMessage">
+                            <input type="text" placeholder="Message" id="message" class="form_control"/>
+                            <div class="input_group_append">
+                                <button id="send" class="btn btn-primary" onclick="send()">보내기</button>
+                                <input type="hidden" value="${chatInfo.buyerId}" id="buyerId"/>
+                                <input type="hidden" value="${chatInfo.usedBookId}" id="usedBookId"/>
+                                <input type="hidden" value="${chatInfo.sellerId}" id="sellerId"/>                       
+                            </div>                    
+                        </div>                
+                    </div>
+                </div>
+        </div>
     
     
     <th:block layout:fragment="script" th:inline="javascript">
